Add support for applying a policy when upgrading a bom’s versions
Previously, when upgrading a bom’s versions all newer versions would
be suggested.

This commit introduces support for upgrade policies which restricts
The suggested versions to those that match the configured policy.
The policy can be configured via a new configuration file that should
be checked in to source control.

Closes gh-4